+++ lang/de/block_exaaichat.php
@@ -46,40 +46,42 @@ $string['downloadfilename'] = 'block_exaaichat_protokolle';
 $string['blocktitle'] = 'Blocktitel';
 
 $string['restrictusage'] = 'Verwendung auf angemeldete Nutzer beschränken';
-$string['restrictusagedesc'] = 'Wenn dieses Kästchen aktiviert ist, können nur angemeldete Benutzer das Chat-Fenster verwenden.';
+$string['restrictusage:desc'] = 'Wenn dieses Kästchen aktiviert ist, können nur angemeldete Benutzer das Chat-Fenster verwenden.';
 $string['apikey'] = 'API-Schlüssel';
-$string['apikeydesc'] = 'Der API-Schlüssel für dein OpenAI-Konto oder dein Azure OpenAI Konto.';
+$string['apikey:desc'] = 'Der API-Schlüssel für dein OpenAI-Konto oder dein Azure OpenAI Konto.';
 $string['type'] = 'API-Typ';
-$string['typedesc'] = 'Der API-Typ, den dieses Plugin verwenden soll.';
+$string['type:desc'] = 'Der API-Typ, den dieses Plugin verwenden soll.';
 $string['logging'] = 'Protokollierung aktivieren';
-$string['loggingdesc'] = 'Wenn diese Einstellung aktiviert ist, werden alle Benutzernachrichten und KI-Antworten protokolliert.';
+$string['logging:desc'] = 'Wenn diese Einstellung aktiviert ist, werden alle Benutzernachrichten und KI-Antworten protokolliert.';
+$string['logging_retention_period'] = 'Protokollaufbewahrungszeitraum (in Tagen)';
+$string['logging_retention_period:desc'] = 'Die Anzahl der Tage, die Protokolleinträge aufbewahrt werden, bevor sie automatisch gelöscht werden. Setze diesen Wert auf 0, um Protokolle unbegrenzt aufzubewahren.';
 
 $string['assistantheading'] = 'Assistent API Einstellungen';
-$string['assistantheadingdesc'] = 'Diese Einstellungen gelten nur für den Assistenten-API-Typ.';
+$string['assistantheading:desc'] = 'Diese Einstellungen gelten nur für den Assistenten-API-Typ.';
 $string['assistant'] = 'Assistent';
-$string['assistantdesc'] = 'Der Standard-Assistent aus deinem OpenAI-Konto, den du für die Generierung der Antworten verwenden möchtest.';
+$string['assistant:desc'] = 'Der Standard-Assistent aus deinem OpenAI-Konto, den du für die Generierung der Antworten verwenden möchtest.';
 $string['noassistants'] = 'Du hast noch keine Assistenten erstellt. Lege zuerst einen <a target="_blank" href="https://platform.openai.com/assistants">in deinem OpenAI-Konto</a> an.';
 $string['persistconvo'] = 'Konversation merken';
-$string['persistconvodesc'] = 'Wenn dieses Kontrollkästchen aktiviert ist, merkt sich der Assistent die Konversation zwischen Seitenaufrufen innerhalb desselben Kontextes. Verschiedene Blockinstanzen (z. B. in einem anderen Kurs) teilen sich keine Konversation.';
+$string['persistconvo:desc'] = 'Wenn dieses Kontrollkästchen aktiviert ist, merkt sich der Assistent die Konversation zwischen Seitenaufrufen innerhalb desselben Kontextes. Verschiedene Blockinstanzen (z. B. in einem anderen Kurs) teilen sich keine Konversation.';
 
 $string['azureheading'] = 'Azure API Einstellungen';
-$string['azureheadingdesc'] = 'Diese Einstellungen gelten nur für den Azure API-Typ.';
+$string['azureheading:desc'] = 'Diese Einstellungen gelten nur für den Azure API-Typ.';
 $string['resourcename'] = 'Ressourcenname';
-$string['resourcenamedesc'] = 'Der Name deiner Azure OpenAI Ressource.';
+$string['resourcename:desc'] = 'Der Name deiner Azure OpenAI Ressource.';
 $string['deploymentid'] = 'Deployment-ID';
-$string['deploymentiddesc'] = 'Der Name des Deployments, den du beim Bereitstellen gewählt hast.';
+$string['deploymentid:desc'] = 'Der Name des Deployments, den du beim Bereitstellen gewählt hast.';
 $string['apiversion'] = 'API-Version';
-$string['apiversiondesc'] = 'Die zu verwendende API-Version im Format JJJJ-MM-TT.';
+$string['apiversion:desc'] = 'Die zu verwendende API-Version im Format JJJJ-MM-TT.';
 $string['chatheading'] = 'Chat API Einstellungen';
-$string['chatheadingdesc'] = 'Diese Einstellungen gelten nur für die Chat-API und die Azure-API Typen.';
+$string['chatheading:desc'] = 'Diese Einstellungen gelten nur für die Chat-API und die Azure-API Typen.';
 $string['prompt'] = 'Prompt';
-$string['promptdesc'] = 'Der Prompt, der der KI vor dem Gespräch gesendet wird.';
+$string['prompt:desc'] = 'Der Prompt, der der KI vor dem Gespräch gesendet wird.';
 $string['assistantname'] = 'Name des Assistenten';
-$string['assistantnamedesc'] = 'Der Name, den die KI intern für sich selbst verwendet. Er erscheint ebenfalls in den Überschriften des Chat-Fensters.';
+$string['assistantname:desc'] = 'Der Name, den die KI intern für sich selbst verwendet. Er erscheint ebenfalls in den Überschriften des Chat-Fensters.';
 $string['username'] = 'Name des Benutzers';
-$string['usernamedesc'] = 'Der Name, den die KI intern für den Benutzer verwendet. Er erscheint ebenfalls in den Überschriften des Chat-Fensters.';
+$string['username:desc'] = 'Der Name, den die KI intern für den Benutzer verwendet. Er erscheint ebenfalls in den Überschriften des Chat-Fensters.';
 $string['sourceoftruth'] = 'Wissensbasis';
-$string['sourceoftruthdesc'] = 'Obwohl die KI sehr leistungsfähig ist, gibt sie – falls sie eine Antwort nicht kennt – eher selbstbewusst eine falsche Information als dass sie die Antwort verweigert. In diesem Textfeld kannst du häufige Fragen und deren Antworten hinterlegen. Format: <pre>Q: Frage 1<br />A: Antwort 1<br /><br />Q: Frage 2<br />A: Antwort 2</pre>
+$string['sourceoftruth:desc'] = 'Obwohl die KI sehr leistungsfähig ist, gibt sie – falls sie eine Antwort nicht kennt – eher selbstbewusst eine falsche Information als dass sie die Antwort verweigert. In diesem Textfeld kannst du häufige Fragen und deren Antworten hinterlegen. Format: <pre>Q: Frage 1<br />A: Antwort 1<br /><br />Q: Frage 2<br />A: Antwort 2</pre>
 Du kannst außerdem Platzhalter wie folgt angeben:
 Mein Name ist {user.fullname}.
 Heute ist {userdate}.
@@ -88,23 +90,23 @@ Möglicher Notenbereich für den Kurs ist {range:coursetotal}.
 ';
 $string['showlabels'] = 'Labels anzeigen';
 $string['advanced'] = 'Erweitert';
-$string['advanceddesc'] = 'Erweiterte Parameter, die an OpenAI gesendet werden. Bitte nur ändern, wenn du genau weißt, was du tust!';
+$string['advanced:desc'] = 'Erweiterte Parameter, die an OpenAI gesendet werden. Bitte nur ändern, wenn du genau weißt, was du tust!';
 $string['allowinstancesettings'] = 'Instanzbezogene Einstellungen';
-$string['allowinstancesettingsdesc'] = 'Diese Einstellung erlaubt es Lehrenden bzw. jedem mit der Berechtigung, einen Block im jeweiligen Kontext hinzuzufügen, instanzspezifische Einstellungen vorzunehmen. Dies kann zu zusätzlichen Kosten führen (z. B. durch Auswahl teurerer Modelle).';
+$string['allowinstancesettings:desc'] = 'Diese Einstellung erlaubt es Lehrenden bzw. jedem mit der Berechtigung, einen Block im jeweiligen Kontext hinzuzufügen, instanzspezifische Einstellungen vorzunehmen. Dies kann zu zusätzlichen Kosten führen (z. B. durch Auswahl teurerer Modelle).';
 $string['model'] = 'Modell';
-$string['modeldesc'] = 'Das Modell, das die Generierung der Antwort übernimmt. Einige Modelle sind auf natürliche Sprache spezialisiert, andere auf Code.';
+$string['model:desc'] = 'Das Modell, das die Generierung der Antwort übernimmt. Einige Modelle sind auf natürliche Sprache spezialisiert, andere auf Code.';
 $string['models'] = 'Modelle';
 $string['models:desc'] = 'Überschreibung der verfügbaren Modelle für die Auswahl im Kurs';
 $string['temperature'] = 'Temperatur';
-$string['temperaturedesc'] = 'Steuert die Zufälligkeit: Ein niedrigerer Wert führt zu weniger zufälligen Antworten. Je näher die Temperatur an 0 liegt, desto deterministischer und repetitiver wird die Ausgabe.';
+$string['temperature:desc'] = 'Steuert die Zufälligkeit: Ein niedrigerer Wert führt zu weniger zufälligen Antworten. Je näher die Temperatur an 0 liegt, desto deterministischer und repetitiver wird die Ausgabe.';
 $string['maxlength'] = 'Maximale Länge';
-$string['maxlengthdesc'] = 'Maximale Anzahl von Tokens, die generiert werden können. Eingaben teilen sich das Token-Limit zwischen Prompt und Antwort. (Ein Token entspricht grob 4 Zeichen normalen englischen Textes)';
+$string['maxlength:desc'] = 'Maximale Anzahl von Tokens, die generiert werden können. Eingaben teilen sich das Token-Limit zwischen Prompt und Antwort. (Ein Token entspricht grob 4 Zeichen normalen englischen Textes)';
 $string['topp'] = 'Top P';
-$string['toppdesc'] = 'Steuert Diversität mittels Nucleus Sampling: 0.5 bedeutet, dass nur die wahrscheinlichsten 50% (gewichtete Optionen) betrachtet werden.';
+$string['topp:desc'] = 'Steuert Diversität mittels Nucleus Sampling: 0.5 bedeutet, dass nur die wahrscheinlichsten 50% (gewichtete Optionen) betrachtet werden.';
 $string['frequency'] = 'Frequenz-Penalty';
-$string['frequencydesc'] = 'Legt fest, wie stark neue Tokens bestraft werden, basierend auf ihrer bisherigen Häufigkeit im Text. Reduziert die Wahrscheinlichkeit, dass identische Zeilen wiederholt werden.';
+$string['frequency:desc'] = 'Legt fest, wie stark neue Tokens bestraft werden, basierend auf ihrer bisherigen Häufigkeit im Text. Reduziert die Wahrscheinlichkeit, dass identische Zeilen wiederholt werden.';
 $string['presence'] = 'Präsenz-Penalty';
-$string['presencedesc'] = 'Legt fest, wie stark neue Tokens bestraft werden, basierend darauf, ob sie bereits im Text vorkommen. Erhöht die Wahrscheinlichkeit, dass neue Themen angesprochen werden.';
+$string['presence:desc'] = 'Legt fest, wie stark neue Tokens bestraft werden, basierend darauf, ob sie bereits im Text vorkommen. Erhöht die Wahrscheinlichkeit, dass neue Themen angesprochen werden.';
 
 $string['config_assistant'] = 'Assistent';
 $string['config_assistant_help'] = 'Wähle den Assistenten, den du für diesen Block verwenden möchtest. Weitere Assistenten können im OpenAI-Konto, das dieser Block nutzt, erstellt werden.';
@@ -184,9 +186,9 @@ $string['vectorstoreids'] = 'Vektorspeicher-IDs';
 
 // Debug file logging setting.
 $string['debugfilelogging'] = 'Debug-Protokollierung aktivieren';
-$string['debugfileloggingdesc'] = 'Alle API-Aufrufe (Benutzernachrichten, KI-Antworten und Funktionsaufrufe) werden in der Datei moodledata/log/exaaichat.log protokolliert.<br/>\nDies kann bei der Fehlersuche hilfreich sein, aber die Protokolldatei kann sehr schnell sehr groß werden. Daher sollte diese Einstellung nur vorübergehend aktiviert werden.';
+$string['debugfilelogging:desc'] = 'Alle API-Aufrufe (Benutzernachrichten, KI-Antworten und Funktionsaufrufe) werden in der Datei moodledata/log/exaaichat.log protokolliert.<br/>\nDies kann bei der Fehlersuche hilfreich sein, aber die Protokolldatei kann sehr schnell sehr groß werden. Daher sollte diese Einstellung nur vorübergehend aktiviert werden.';
 $string['additionalmessage'] = 'Zusätzlicher Text für jede Nachricht';
-$string['additionalmessagedesc'] = 'Dieser Text wird vor dem Senden an die KI an jede Benutzernachricht angehängt.';
+$string['additionalmessage:desc'] = 'Dieser Text wird vor dem Senden an die KI an jede Benutzernachricht angehängt.';
 
 // Type select option labels.
 $string['type_chat'] = 'OpenAI: Chat API';